refactor: move getEventAdminAreas to event area service
